feat/mux-player-react: cleanup/split into modules, including type definition module.
