+++ packages/mux-player-react/src/components/Spacer.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+const Spacer = () => {
+  return (
+    <div
+      style={{
+        flexGrow: 1,
+        height: "100%",
+        backgroundColor: "var(--media-control-background, rgba(20,20,30, 0.7))",
+      }}
+    ></div>
+  );
+};
+
+export default Spacer;
