+++ packages/mux-player-react/src/constants.ts
@@ -1,0 +1,4 @@
+export const MediaChromeSizes: { LG: "large"; SM: "small" } = {
+  LG: "large",
+  SM: "small",
+};
