+++ packages/mux-player-react/src/index.tsx
@@ -1,417 +1,21 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
-import * as CSS from "csstype";
+import React, { useCallback, useEffect, useState } from "react";
 import MuxVideo from "@mux-elements/mux-video-react";
-import type { Props as MuxVideoProps } from "@mux-elements/mux-video-react";
-import {
-  MediaController,
-  MediaControlBar,
-  MediaPlayButton,
-  MediaSeekBackwardButton,
-  MediaSeekForwardButton,
-  MediaMuteButton,
-  MediaVolumeRange,
-  MediaTimeRange,
-  MediaTimeDisplay,
-  MediaCaptionsButton,
-  MediaPlaybackRateButton,
-  MediaPipButton,
-  MediaFullscreenButton,
-} from "./media-chrome";
+import { MediaController } from "./media-chrome";
 import { useRef } from "react";
-import Loading from "./media-chrome/components/loading";
-import AirPlayButton from "./media-chrome/components/air-play-button";
-import { StreamTypes } from "@mux-elements/playback-core";
-import { useTimeoutWhen } from "./useTimeoutWhen";
-import { useBoundingclientrect } from "./useBoundingclientrect";
-import { useCombinedRefs } from "./useCombinedRefs";
+import { StreamTypes, ValueOf } from "@mux-elements/playback-core";
+import { useBoundingclientrect } from "./hooks/useBoundingclientrect";
+import { useCombinedRefs } from "./hooks/useCombinedRefs";
+import DefaultChromeRenderer from "./components/DefaultChromeRenderer";
+import { MediaChromeSizes } from "./constants";
+import { MuxPlayerProps } from "./types";
+import {
+  getChromeStylesFromProps,
+  getPosterURLFromPlaybackId,
+  getStoryboardURLFromPlaybackId,
+  hasVolumeSupportAsync,
+} from "./utils";
 
 export { StreamTypes };
-
-const MediaChromeSizes = {
-  LG: "large",
-  SM: "small",
-};
-
-declare global {
-  interface Window {
-    WebKitPlaybackTargetAvailabilityEvent?: Event;
-  }
-
-  interface HTMLVideoElement {
-    webkitShowPlaybackTargetPicker?: Function;
-  }
-}
-
-declare module "csstype" {
-  interface Properties {
-    "--media-background-color"?: CSS.Properties["backgroundColor"];
-    "--media-control-background"?: CSS.Properties["backgroundColor"];
-    "--media-control-hover-background"?: CSS.Properties["backgroundColor"];
-    "--media-button-icon-width"?: CSS.Properties["width"];
-    "-webkit-transform"?: CSS.Properties["transform"];
-    "-ms-transform"?: CSS.Properties["transform"];
-    // ...or allow any other property
-    // [index: string]: any;
-  }
-}
-
-const getPosterURLFromPlaybackId = (playbackId: MuxVideoProps["playbackId"]) =>
-  `https://image.mux.com/${playbackId}/thumbnail.jpg`;
-
-const getStoryboardURLFromPlaybackId = (
-  playbackId: MuxVideoProps["playbackId"]
-) => `https://image.mux.com/${playbackId}/storyboard.vtt`;
-
-const Spacer = () => {
-  return (
-    <div
-      style={{
-        flexGrow: 1,
-        height: "100%",
-        backgroundColor: "var(--media-control-background, rgba(20,20,30, 0.7))",
-      }}
-    ></div>
-  );
-};
-
-let testMediaEl: HTMLMediaElement | undefined;
-const getTestMediaEl = (nodeName = "video") => {
-  if (testMediaEl) return testMediaEl;
-  if (typeof window !== "undefined") {
-    testMediaEl = document.createElement(nodeName as "video" | "audio");
-  }
-  return testMediaEl;
-};
-
-const hasVolumeSupportAsync = async (
-  mediaEl: HTMLMediaElement | undefined = getTestMediaEl()
-) => {
-  if (!mediaEl) return false;
-  const prevVolume = mediaEl.volume;
-  mediaEl.volume = prevVolume / 2 + 0.1;
-  return new Promise<boolean>((resolve, reject) => {
-    setTimeout(() => {
-      resolve(mediaEl.volume !== prevVolume);
-    }, 0);
-  });
-};
-
-type ChromeProps = {
-  onAirPlaySelected?: React.MouseEventHandler;
-  supportsAirPlay?: boolean;
-  supportsVolume?: boolean;
-  loading?: boolean;
-  paused?: boolean;
-  captionsAvailable?: boolean;
-  streamType?: MuxVideoProps["streamType"];
-  playerSize?: string;
-};
-
-export const VodChromeSmall: React.FC<ChromeProps> = (props) => {
-  const {
-    onAirPlaySelected,
-    supportsAirPlay = false,
-    supportsVolume = false,
-    captionsAvailable = false,
-    loading = false,
-    paused,
-  } = props;
-  const [showLoading, setShowLoading] = useState(loading && !paused);
-  useTimeoutWhen(
-    () => {
-      setShowLoading(loading && !paused);
-    },
-    500,
-    loading && !paused
-  );
-  useEffect(() => {
-    const nextShowLoading = loading && !paused;
-    if (nextShowLoading) return;
-    setShowLoading(nextShowLoading);
-  }, [loading, paused]);
-  return (
-    <>
-      <div
-        slot="centered-chrome"
-        style={{
-          "--media-background-color": "none",
-          "--media-control-hover-background": "none",
-          "--media-control-background": "none",
-          "--media-button-icon-width": "100%",
-          width: "100%",
-          display: "flex",
-          flexFlow: "row",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
-        {showLoading ? (
-          <Loading />
-        ) : (
-          <>
-            <MediaSeekBackwardButton
-              style={{ padding: 0, width: "20%" }}
-            ></MediaSeekBackwardButton>
-            <MediaPlayButton
-              style={{ padding: 0, width: "20%" }}
-            ></MediaPlayButton>
-            <MediaSeekForwardButton
-              style={{ padding: 0, width: "20%" }}
-            ></MediaSeekForwardButton>
-          </>
-        )}
-      </div>
-      <MediaControlBar>
-        <MediaTimeRange></MediaTimeRange>
-        <MediaTimeDisplay show-duration></MediaTimeDisplay>
-      </MediaControlBar>
-      <MediaControlBar>
-        <MediaMuteButton></MediaMuteButton>
-        {supportsVolume && <MediaVolumeRange></MediaVolumeRange>}
-        <Spacer />
-        {captionsAvailable && <MediaCaptionsButton></MediaCaptionsButton>}
-        <MediaPipButton></MediaPipButton>
-        <MediaFullscreenButton></MediaFullscreenButton>
-        {supportsAirPlay && <AirPlayButton onClick={onAirPlaySelected} />}
-      </MediaControlBar>
-    </>
-  );
-};
-
-export const VodChromeLarge: React.FC<ChromeProps> = (props) => {
-  const {
-    onAirPlaySelected,
-    supportsAirPlay = false,
-    supportsVolume = false,
-    captionsAvailable = false,
-    loading = false,
-    paused,
-  } = props;
-  const [showLoading, setShowLoading] = useState(loading && !paused);
-  useTimeoutWhen(
-    () => {
-      setShowLoading(loading && !paused);
-    },
-    500,
-    loading && !paused
-  );
-  useEffect(() => {
-    const nextShowLoading = loading && !paused;
-    if (nextShowLoading) return;
-    setShowLoading(nextShowLoading);
-  }, [loading, paused]);
-  return (
-    <>
-      <div
-        slot="centered-chrome"
-        style={{
-          "--media-background-color": "none",
-          "--media-control-hover-background": "none",
-          "--media-control-background": "none",
-          "--media-button-icon-width": "100%",
-          width: "100%",
-          display: "flex",
-          flexFlow: "row",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
-        {showLoading ? <Loading /> : <></>}
-      </div>
-      <MediaControlBar>
-        <MediaPlayButton></MediaPlayButton>
-        <MediaSeekForwardButton></MediaSeekForwardButton>
-        <MediaSeekBackwardButton></MediaSeekBackwardButton>
-        <MediaMuteButton></MediaMuteButton>
-        {supportsVolume && <MediaVolumeRange></MediaVolumeRange>}
-        <MediaTimeRange></MediaTimeRange>
-        <MediaTimeDisplay
-          show-duration
-          style={{ color: "inherit" }}
-        ></MediaTimeDisplay>
-        {captionsAvailable && <MediaCaptionsButton></MediaCaptionsButton>}
-        <MediaPipButton></MediaPipButton>
-        <MediaFullscreenButton></MediaFullscreenButton>
-        {supportsAirPlay && <AirPlayButton onClick={onAirPlaySelected} />}
-      </MediaControlBar>
-    </>
-  );
-};
-
-export const LiveChromeSmall: React.FC<ChromeProps> = (props) => {
-  const {
-    onAirPlaySelected,
-    supportsAirPlay = false,
-    supportsVolume = false,
-    captionsAvailable = false,
-    loading = false,
-    paused,
-  } = props;
-  const [showLoading, setShowLoading] = useState(loading && !paused);
-  useTimeoutWhen(
-    () => {
-      setShowLoading(loading && !paused);
-    },
-    500,
-    loading && !paused
-  );
-  useEffect(() => {
-    const nextShowLoading = loading && !paused;
-    if (nextShowLoading) return;
-    setShowLoading(nextShowLoading);
-  }, [loading, paused]);
-  return (
-    <>
-      <div
-        slot="centered-chrome"
-        style={{
-          "--media-background-color": "none",
-          "--media-control-hover-background": "none",
-          "--media-control-background": "none",
-          "--media-button-icon-width": "100%",
-          width: "100%",
-          display: "flex",
-          flexFlow: "row",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
-        {showLoading ? (
-          <Loading />
-        ) : (
-          <>
-            <MediaPlayButton
-              style={{ padding: 0, width: "20%" }}
-            ></MediaPlayButton>
-          </>
-        )}
-      </div>
-      <MediaControlBar>
-        <MediaMuteButton></MediaMuteButton>
-        {supportsVolume && <MediaVolumeRange></MediaVolumeRange>}
-        <Spacer />
-        {captionsAvailable && <MediaCaptionsButton></MediaCaptionsButton>}
-        <MediaPipButton></MediaPipButton>
-        <MediaFullscreenButton></MediaFullscreenButton>
-        {supportsAirPlay && <AirPlayButton onClick={onAirPlaySelected} />}
-      </MediaControlBar>
-    </>
-  );
-};
-export const LiveChromeLarge: React.FC<ChromeProps> = (props) => {
-  const {
-    onAirPlaySelected,
-    supportsAirPlay = false,
-    supportsVolume = false,
-    captionsAvailable = false,
-    loading = false,
-    paused,
-  } = props;
-  const [showLoading, setShowLoading] = useState(loading && !paused);
-  useTimeoutWhen(
-    () => {
-      setShowLoading(loading && !paused);
-    },
-    500,
-    loading && !paused
-  );
-  useEffect(() => {
-    const nextShowLoading = loading && !paused;
-    if (nextShowLoading) return;
-    setShowLoading(nextShowLoading);
-  }, [loading, paused]);
-  return (
-    <>
-      <div
-        slot="centered-chrome"
-        style={{
-          "--media-background-color": "none",
-          "--media-control-hover-background": "none",
-          "--media-control-background": "none",
-          "--media-button-icon-width": "100%",
-          width: "100%",
-          display: "flex",
-          flexFlow: "row",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
-        {showLoading ? <Loading /> : <></>}
-      </div>
-      <MediaControlBar>
-        <MediaPlayButton></MediaPlayButton>
-        <MediaMuteButton></MediaMuteButton>
-        {supportsVolume && <MediaVolumeRange></MediaVolumeRange>}
-        <Spacer />
-        {captionsAvailable && <MediaCaptionsButton></MediaCaptionsButton>}
-        <MediaPlaybackRateButton></MediaPlaybackRateButton>
-        <MediaPipButton></MediaPipButton>
-        <MediaFullscreenButton></MediaFullscreenButton>
-        {supportsAirPlay && <AirPlayButton onClick={onAirPlaySelected} />}
-      </MediaControlBar>
-    </>
-  );
-};
-
-export const DefaultChromeRenderer: React.FC<ChromeProps> = (props) => {
-  const { streamType, playerSize } = props;
-  if (playerSize === MediaChromeSizes.SM) {
-    if (streamType === StreamTypes.LIVE || streamType === StreamTypes.LL_LIVE) {
-      return <LiveChromeSmall {...props} />;
-    }
-    return <VodChromeSmall {...props} />;
-  }
-  if (streamType === StreamTypes.LIVE || streamType === StreamTypes.LL_LIVE) {
-    return <LiveChromeLarge {...props} />;
-  }
-  return <VodChromeLarge {...props} />;
-};
-
-type ReactInstanceBasic = React.ReactElement | null;
-export type MuxPlayerProps = Partial<MuxVideoProps> & {
-  onPlayerReady?: () => void; // params?
-  onError?: (e: ErrorEvent) => void;
-  defaultShowCaptions?: boolean;
-  ChromeRenderer?: (props: ChromeProps) => ReactInstanceBasic;
-  primaryColor?: React.CSSProperties["color"];
-  secondaryColor?: React.CSSProperties["color"];
-  tertiaryColor?: React.CSSProperties["color"];
-};
-
-const getChromeStylesFromProps = (props: MuxPlayerProps) => {
-  const { primaryColor, secondaryColor, tertiaryColor } = props;
-
-  const primaryColorStyles = primaryColor
-    ? {
-        "--media-icon-color": primaryColor,
-        "--media-range-thumb-background": primaryColor,
-        "--media-range-bar-color": primaryColor,
-        color: primaryColor,
-      }
-    : {};
-
-  const secondaryColorStyles = secondaryColor
-    ? {
-        "--media-background-color": secondaryColor,
-        "--media-control-hover-background": secondaryColor,
-        "--media-control-background": secondaryColor,
-      }
-    : {};
-
-  const tertiaryColorStyles = tertiaryColor
-    ? {
-        "--media-range-track-background": tertiaryColor,
-      }
-    : {};
-
-  return {
-    maxWidth: "100%",
-    color: "#ffffff",
-    ...primaryColorStyles,
-    ...secondaryColorStyles,
-    ...tertiaryColorStyles,
-  };
-};
 
 const SMALL_BREAKPOINT = 700;
 
@@ -478,7 +82,6 @@ export const MuxPlayer = React.forwardRef<
   const [captionsAvailable, setCaptionsAvailable] = useState<boolean>();
 
   useEffect(() => {
-    console.log("playbackId", playbackId);
     setSrcLoaded(undefined);
     setCaptionsAvailable(undefined);
     setPlayerReady(false);
@@ -586,7 +189,9 @@ export const MuxPlayer = React.forwardRef<
     );
   }, []);
 
-  const [playerSize, setPlayerSize] = useState(MediaChromeSizes.LG);
+  const [playerSize, setPlayerSize] = useState<
+    ValueOf<typeof MediaChromeSizes>
+  >(MediaChromeSizes.LG);
 
   const mediaControllerRef = useRef(null);
   const mediaControllerRect = useBoundingclientrect(mediaControllerRef);
@@ -601,13 +206,6 @@ export const MuxPlayer = React.forwardRef<
     setPlayerSize(nextPlayerSize);
   }, [mediaControllerRect]);
 
-  // Add appropriate conditions for "playerReady" here
-  // * media loaded (?)
-  // * captions availability known
-  // * volume support known
-  // * airplay support known
-  // * bounding rect/size known (?)
-  // * (stretch) aspect ratio resizing done
   const [playerReady, setPlayerReady] = useState(false);
   useEffect(() => {
     const allReady = ![
